# backend-foryou/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ForYou API Documentation",
        "description": "API documentation for the ForYou project",
        "contact": {
            "email": "admin@foryou.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/visitas": {
            "get": {
                "tags": [
                    "Visitas"
                ],
                "summary": "Get list of visitas",
                "description": "Returns list of visitas",
                "operationId": "getVisitas",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Visita"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Visitas"
                ],
                "summary": "Create a new visita",
                "description": "Creates a new visita",
                "operationId": "createVisita",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/VisitaInput"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Visita created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Visita"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/visitas/{id}": {
            "get": {
                "tags": [
                    "Visitas"
                ],
                "summary": "Get visita by ID",
                "description": "Returns a single visita",
                "operationId": "getVisitaById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of visita to return",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Visita"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Visita not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Visitas"
                ],
                "summary": "Update an existing visita",
                "description": "Updates a visita",
                "operationId": "updateVisita",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of visita to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/VisitaInput"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Visita updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Visita"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Visita not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Visitas"
                ],
                "summary": "Delete a visita",
                "description": "Deletes a single visita",
                "operationId": "deleteVisita",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of visita to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "404": {
                        "description": "Visita not found"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "Visita": {
                "title": "Visita",
                "required": [
                    "id_empleado",
                    "id_solicitud",
                    "fecha"
                ],
                "properties": {
                    "id_visita": {
                        "description": "ID de la visita",
                        "type": "integer",
                        "format": "int64"
                    },
                    "id_empleado": {
                        "description": "ID del empleado que realiza la visita",
                        "type": "integer",
                        "format": "int64"
                    },
                    "id_solicitud": {
                        "description": "ID de la solicitud visitada",
                        "type": "integer",
                        "format": "int64"
                    },
                    "fecha": {
                        "description": "Fecha y hora de la visita",
                        "type": "string",
                        "format": "date-time"
                    },
                    "observaciones": {
                        "description": "Observaciones de la visita",
                        "type": "string"
                    },
                    "created_at": {
                        "description": "Fecha de creación",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Fecha de actualización",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "VisitaInput": {
                "title": "Visita Input",
                "required": [
                    "id_empleado",
                    "id_solicitud",
                    "fecha"
                ],
                "properties": {
                    "id_empleado": {
                        "description": "ID del empleado que realiza la visita",
                        "type": "integer",
                        "format": "int64"
                    },
                    "id_solicitud": {
                        "description": "ID de la solicitud visitada",
                        "type": "integer",
                        "format": "int64"
                    },
                    "fecha": {
                        "description": "Fecha y hora de la visita",
                        "type": "string",
                        "format": "date-time"
                    },
                    "observaciones": {
                        "description": "Observaciones de la visita",
                        "type": "string"
                    }
                },
                "type": "object"
            }
        }
    },
...
}